+++ src/components/EventEditor/FormComponents/PlaceComponent/PlaceComponent.js
@@ -1,11 +1,14 @@
-import {Form, Input} from 'antd';
 import React from 'react';
+import {Form, Input} from 'antd';
+import GoogleMapPlaceInput from '../../../GoogleMapPlaceInput';
+
 import './place-component.css';
 
 const PlaceComponent = () => {
   return (
     <Form.Item name="place" label="Place">
-      <Input name="place" style={{width: '100%'}} placeholder="Please enter place" />
+      {/* <Input name="place" style={{width: '100%'}} placeholder="Please enter place" /> */}
+      <GoogleMapPlaceInput />
     </Form.Item>
   );
 };
