+++ src/components/GoogleMapPlaceInput/GoogleMapPlaceInput.css
@@ -1,0 +1,4 @@
+.input-map {
+  width: 100%;
+  margin-bottom: 8px;
+}
