+++ src/components/GoogleMapPlaceInput/index.js
@@ -1,0 +1,3 @@
+import GoogleMapPlaceInput from './GoogleMapPlaceInput';
+
+export default GoogleMapPlaceInput;
