+++ src/components/Header/HeaderToolsPanel/HeaderToolsPanel.js
@@ -10,7 +10,6 @@ import {Button, Dropdown, Menu, Space, Tooltip} from 'antd';
 import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {switchVisibilityHiddenEvents} from '../../../actions';
-import InputGoogleMap from '../../InputGoogleMap';
 import NewEventButton from '../../NewEventButton';
 import TimeZoneSelect from '../../TimeZoneSelect';
 import ScheduleViewSelect from '../ScheduleViewSelect';
@@ -67,8 +66,6 @@ const HeaderToolsPanel = () => {
           <TimeZoneSelect />
         </Tooltip>
       </Space>
-
-      <InputGoogleMap />
 
       <Space>
         <ScheduleViewSelect />
