--- src/components/InputGoogleMap/index.js
@@ -1,3 +1,0 @@
-import InputGoogleMap from './InputGoogleMap';
-
-export default InputGoogleMap;
